backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Science Core
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# Advanced Encoders & Balancing
category-encoders>=2.6.0
imbalanced-learn>=0.11.0

# Validation
pandera>=0.17.0

# Visualization
plotly>=5.18.0
seaborn>=0.13.0
matplotlib>=3.8.0

# File I/O
pyarrow>=14.0.0
openpyxl>=3.1.0

# NLP (Phase 12)
nltk>=3.8.0
gensim>=4.3.0
vaderSentiment==3.3.2

# BI Layer - Complete
openai==1.35.0     # For GPT-4 API
google-generativeai==0.8.3  # For Gemini API
plotly==5.18.0
kaleido==0.2.1  # For static plot export
scipy==1.11.4  # For advanced stats

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
